requirements.txt
pytest
pytest-asyncio
playwright
allure-pytest